more changes to reduce diff. TODO: adjust tests
