+++ q2_types/reference_db/tests/test_format.py
@@ -5,17 +5,13 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
-import tempfile
-import shutil
-import os
 from qiime2.plugin.testing import TestPluginBase
 from q2_types.reference_db._format import (
-    DiamondDatabaseFileFmt, DiamondDatabaseDirFmt, EggnogRefBinFileFmt,
-    EggnogRefDirFmt, NCBITaxonomyNamesFormat, NCBITaxonomyNodesFormat,
-    NCBITaxonomyDirFmt, NCBITaxonomyBinaryFileFmt,
-    EggnogProteinSequencesDirFmt, EggnogRefTextFileFmt, HmmerDirFmt,
-    HmmerIdmapFileFmt
-)
+        DiamondDatabaseFileFmt, DiamondDatabaseDirFmt, EggnogRefBinFileFmt,
+        EggnogRefDirFmt, NCBITaxonomyNamesFormat, NCBITaxonomyNodesFormat,
+        NCBITaxonomyDirFmt, NCBITaxonomyBinaryFileFmt,
+        EggnogProteinSequencesDirFmt, EggnogRefTextFileFmt
+        )
 from qiime2.plugin import ValidationError
 
 
@@ -154,78 +150,6 @@ class TestRefFormats(TestPluginBase):
             r"Invalid line at line 9"
         ):
             fmt_obj.validate()
-
-    def test_HmmerDirFmt_valid(self):
-        fmt = HmmerDirFmt(self.get_data_path("hmmer/bacteria"), 'r')
-        fmt.validate()
-
-    def test_HmmerDirFmt_invalid_idmap_1(self):
-        fmt = HmmerIdmapFileFmt(self.get_data_path(
-            "hmmer/invalid_idmaps/1.hmm.idmap"), 'r'
-        )
-        with self.assertRaisesRegex(
-            ValidationError,
-            "Expected index and an alphanumeric code separated "
-            "by a single space."
-        ):
-            fmt.validate(level="min")
-
-    def test_HmmerDirFmt_invalid_idmap_2(self):
-        fmt = HmmerIdmapFileFmt(self.get_data_path(
-            "hmmer/invalid_idmaps/2.hmm.idmap"), 'r'
-        )
-        with self.assertRaisesRegex(
-            ValidationError,
-            "Expected index and an alphanumeric code separated "
-            "by a single space."
-        ):
-            fmt.validate(level="min")
-
-    def test_HmmerDirFmt_invalid_idmap_3(self):
-        fmt = HmmerIdmapFileFmt(self.get_data_path(
-            "hmmer/invalid_idmaps/3.hmm.idmap"), 'r'
-        )
-        with self.assertRaisesRegex(
-            ValidationError,
-            'Expected index'
-        ):
-            fmt.validate(level="min")
-
-    def test_HmmerDirFmt_invalid_idmap_4(self):
-        fmt = HmmerIdmapFileFmt(self.get_data_path(
-            "hmmer/invalid_idmaps/4.hmm.idmap"), 'r'
-        )
-        with self.assertRaisesRegex(
-            ValidationError,
-            "Expected index and an alphanumeric code separated "
-            "by a single space."
-        ):
-            fmt.validate(level="min")
-
-    def test_HmmerDirFmt_missing_hmm(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            shutil.copytree(
-                self.get_data_path("hmmer/bacteria"), tmp, dirs_exist_ok=True
-            )
-            os.remove(f"{tmp}/bacteria.hmm.h3f")
-            fmt = HmmerDirFmt(tmp, 'r')
-            with self.assertRaisesRegex(
-                ValidationError, "Missing one or more files"
-            ):
-                fmt.validate(level="min")
-
-    def test_HmmerDirFmt_missing_fa(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            shutil.copytree(
-                self.get_data_path("hmmer/bacteria"), tmp, dirs_exist_ok=True
-            )
-            for file in ["a", "b", "b2"]:
-                os.remove(f"{tmp}/{file}.fa")
-            fmt = HmmerDirFmt(tmp, 'r')
-            with self.assertRaisesRegex(
-                ValidationError, "Missing one or more files"
-            ):
-                fmt.validate(level="min")
 
 
 class TestNCBIFormats(TestPluginBase):
